LPD-56206 Move logic to detect small resolution to new hook
